pyproject.toml
[project]
name = "game-data-orchestration"
version = "0.1.0"
description = ""
authors = [
    {name = "Kerim Tricic",email = "kerim.tricic@dualsharks.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel (>=7.1.0,<8.0.0)",
    "pandas (>=2.2.0,<3.0.0)",
    "pyarrow (>=17.0.0,<18.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "httpx (>=0.25.0,<0.28.0)",
    "luaparser (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false